[FEAT] 게시글 좋아요 기능 연동, API 수정

+++ src/main/java/com/example/pintbackend/dto/postDto/LikeResponse.java
@@ -12,5 +12,5 @@ package com.example.pintbackend.dto.postDto;
 
 
 public record LikeResponse(boolean isLiked,
-                           long likeCount) {
+                           int likeCount) {
 }
